add demos
